+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [UNRELEASED]
 
 - parsing <think> tags as reasoning part of the response
-- added more color themes
+- implemented better support for themes + new themes added
 
 ## [0.26.0]